+++ stores/dashboard/forms.py
@@ -1,7 +1,5 @@
 from django import VERSION, forms
-from django.conf import settings
 from django.contrib.gis.forms import fields
-from django.contrib.gis.geoip2 import HAS_GEOIP2
 from django.db.models import Q
 from django.forms import models as modelforms
 from django.utils.encoding import force_text
@@ -33,19 +31,12 @@ class StoreForm(forms.ModelForm):
         }
 
     def __init__(self, *args, **kwargs):
-        current_ip = kwargs.pop('current_ip', None)
-        super(StoreForm, self).__init__(*args, **kwargs)
-
+        super().__init__(*args, **kwargs)
         # Make sure that we store the initial data as GeoJSON so that
         # it is easier for us to use it in Javascript.
         instance = kwargs.get('instance', None)
         if instance:
             self.initial['location'] = instance.location.geojson
-        elif HAS_GEOIP2 and getattr(settings, 'GEOIP_ENABLED', True):
-            from django.contrib.gis.geoip2 import GeoIP2
-            point = GeoIP2().geos(current_ip)
-            if point:
-                self.initial['location'] = point.geojson
 
     def clean_reference(self):
         ref = self.cleaned_data['reference']